Extend Split to work with multiple separators, replace boost::split with SplitString
Summary:
This removes the remaining occurrences of boost::split and replaces them with our own SplitString. To be able to do so, this extends the function spanparsing::Split to work with multiple separators. Finally this removes 3 more files from lint-includes.py.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/25057 | core#25057]]

Test Plan: `ninja all check-all bitcoin-fuzzers`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19741