[avalanche] Fix logging of vote statuses
Summary:
Fallthrough is causing all vote statuses to show "finalized" in the logs.

This patch fixes it and actually checks the logs for a status other than finalized in a test.

Test Plan:
This can be run with only the patch to the test to see that it fails or with the full patch
to see the issue is now fixed:
  ninja bitcoind && ./test/functional/test_runner.py abc_p2p_avalanche_voting.py

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11452