[avalanche] Use the ProofComparator for sets and maps
Summary: There are places where the comparator is not important as long as the ordering is consistent, and the default comparator is used. This assumes that the key actually has a default comparator. This diff adds a simple comparator and makes explicit use of it where possible so updating the key type will not be an issue as long as it is managed by the comparator. There is no change in behavior.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11463