[avalanche] Remove unnecessary code path in addNodeToQuorum
Summary:
This NO_NODE code path is in fact not helpful for testing since it silently
breaks accounting of successfulVotes. Instead of writing tests to test the
test code, we can just get rid of it. Now the voterecord_tests test what we
expect them to.

Depends on D11500

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11501