refactor: Remove defunct attributes.h includes
Summary:
Since the removal of NODISCARD the only attributes def is LIFETIMEBOUND, and it's included in many more places that it is used.

This removes all includes which do not have an associated use of LIFETIMEBOUND, and adds it to  src/validationinterface.cpp.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/25184 | core#25184]]

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13266