Fix an exception message in blockencodings_tests when the compact size MAX_SIZE is not the limiting factor
Summary:
The TransactionRequest deserialization overflow test was ported in D5049 and adapted to fit our codebase and check the actual exception message. This include a test for an actually unreachable test case, which is added as a safety net in case the MAX_SIZE constant is bumped.

However after a refactor backported in D8790, this exception message is no longer the same as the overflow is now managed by the  DifferenceFormatter. This diff fixes the error message.

Test Plan:
  ninja check

Edit serialize.h to bump MAX_SIZE to 0xFFFFFFFE, then
  ninja check

Reviewers: #bitcoin_abc, sdulfari, PiRK

Reviewed By: #bitcoin_abc, sdulfari, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11517