[Cashtab] Hide Add To Contacts option if sender is a contact
Summary: Related to T2412. Added conditional logic to hide the Add To Contact option in TxHistory if the sender is already a contact.

Test Plan:
cd web/cashtab && npm start
From TxHistory:
Click on a received Tx from a wallet that is not a contact and add the contact
Refresh the page and navigate back to TxHistory
Observe that the Add to Contacts option no longer appears for Txs received from the contact

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11525