[Cashtab] [Chronik Websocket Notifications] [Part 7] Display XEC notifications from ws msg
Summary:
T2447

Depends on D11537

Display notification when websocket picks up an incoming XEC tx

NB: As this is part of a stacked diff, it is expected behavior for the app to notify twice for incoming XEC tx. This will be corrected in a later diff on this stack.

Test Plan:
`npm start`
Send yourself an XEC tx
Note a tx that reads "eCash Received" -- this is from the ws
Change wallets and repeat

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D11538