[avalanche] Add a way to get the lowest scoring proof in a pool
Summary:
This is the next piece needed in order to cap the size of proof pools by the
number of proofs. The PeerManager can use this to kick out the lowest scoring
proof and replace it with a better candidate so the validProofPool always
contains the best limited set of proofs. This can be used in a similar way to
limit the orphan pool.

Depends on D11561

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11562