[avalanche] Respond to missing proof indices request
Summary:
Retrieve missing proofs by matching the requested indices against the stored proof radix tree for the requesting peer. The proofs are sent with individual avaproof messages so we don't have to care about the message being oversized. The radix tree is nulled after this so we only process the requests one at a time.

Depends on D11545.

Test Plan:
  ninja check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11575