drop network.{h|cpp}
Summary:
This contained only a single constant which was used in a single callsite.

More simplifications could be done around the now unneccery `networks/abc/*` directory structure, but this should be done in a separate diff as it requires modifying scripts and deciding whether we want to be ready to support multiple networks again or not.

Depends on D11563

Test Plan:
```
cmake .. -GNinja
ninja all check-all
```

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11564