[Cashtab] Replaced all instances of react-copy-to-clipboard
Summary: Related to T2291. In order to minimize reliance on external libraries, react-copy-to-clipboard has been deprecated throughout the app. However, the package remains in the package-lock.json file as it is a dependency of antd.

Test Plan:
cd web/cashtab && npm start

run `grep -r "import { CopyToClipboard } from 'react-copy-to-clipboard';" src/` and observe that nothing is returned

from the home screen:
use the copy txid button
use the copy msg button

navigate to the receive screen
click the QR code

navigate to Send.js
generate a message signature and click on the message signature input

for each, observe that the correct data is copied to clipboard and that the success notification fires

Reviewers: bytesofman, emack, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11579