[avalanche] Cleanup unrequested radix tree after a timeout
Summary:
This ensures that the radix tree don't add up too much if the compact proofs are never requested.
The cleanup is done in the same thread than the periodic getavaaddr requests in order to reuse some of the code. The thread will later also be extended with getavaproofs requests so this is renamed appropriately.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11612