[avalanche] Use the number nodes from which we've received avaproofs as a criterion for quorum readiness
Summary:
This maintains a count of the successful compact proofs exchanges and enforce a minimum quantity of node that sent avaproofs message before the quorum is considered valid.
This is a best effort to give confidence that our proof set is in par with the other nodes from the network before we start polling.
The threshold is put to 8: because we need at least 8 nodes in our quorum to have converging votes there is no point using a lower value. This means that 1 out of 8 outbounds should be honest for us to get a good set of proofs.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: Mengerian, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11601