test: Remove MiniWallet mempool_valid option
Summary:
```
It seems an unnecessary burden to force MiniWallet call-sites to figure out for each tx whether it is mempool valid or not. The result will only be used for internal sanity checks. So remove the option:
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/25356 | core#25356]].

Depends on D12732.

Note: most tests are not converted to MiniWallet yet, but this change is trivial enough that it's not a problem.

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12733