[Cashtab] Patch bracket color for Documentation link
Summary: Related to T2496. Brackets around documentation link should be grey, not black.

Test Plan:
cd web/cashtab && npm start
navigate to Settings tab and scroll to the bottom of the page
observe that the brackets around the documentation link are no longer black

used lightWhite because it looks a little better than lightGrey, but can change it if lightGrey is preferred

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11620