[avalanche] Don't process proofs that are not worth polling
Summary:
It is possible for a proof to turn orphan while being voted on if a block is connected in the meantime. In this case it becomes pointless to keep polling for it. More generally it's more robust to also discard polls for proofs that get dropped, even if there should be no such case.

Depends on D11641.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix, sdulfari

Reviewed By: #bitcoin_abc, deadalnix, sdulfari

Subscribers: sdulfari, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11630