[avalanche] Make IsWorthPolling a private method of the processor
Summary:
This is in preparation for adding a similar method for proofs that will need to access the processor internals. Using a private method for the blocks will allow for calling it in lambda for both blocks and proofs and so share  the code. There is no change in behavior. Extracted from D11630.

Depends on D11640.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11641