[avalanche] Cleanup dangling proofs
Summary:
This removes the proofs attached to dangling peers, i.e. the peers with no node that have been registered for some time. This means that we will discard the proof if no node claimed it or if they disconnected. This will also sink the conflicting pool as the proofs will be pulled back to valid, and eventually cleanup up the same way if no node attached to it as well.
The callsite for this method will be added in a follow up with the associated functional tests.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11659