[avalanche] Use immature UTXO to orphan the proof in known_score_tracking test
Summary:
Since D11622, orphans can have valid UTXOs but not be mature enough to be valid.
The next major step is to limit the orphan pool to only immature proofs and not
accept proofs with missing UTXOs.

This patch prepares us for that change by migrating one of the instances of a
peer manager test that uses an orphan proof with a missing UTXO to use an
immature one instead. This will make review of the the future immature-only
orphan pool patch simpler since this change is self-contained.

Depends on D11622 and D11664

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11665