[avalanche] Simplify the proof reconciliation logic
Summary: This diff moves some of the logic surrounding `addProofToReconcile` and deduplicate some other. The `isWorthPolling` method now takes care of determining if the proof needs to be voted on and we can leverage this to simplify the code in the network layer. It also makes it easier to use the addProofToReconcile method.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11672