[avalanche] Skip processing compact proofs with no shortid
Summary:
A compact proofs message (`avaproofs`) with an empty shortid list is actually processed, accounted as a valid message and an empty request is sent in response. We can simplify this logic by skipping all this if there is nothing to process, not sending the empty request and not counting the message as a valid one. As a consequence this also improves the compact proof node counter logic (we are sure we received a useful compact proofs message vs any compact proofs message).

Depends on D11680.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11681