[avalanche] Request addresses and proofs from our inbounds while the quorum is not established
Summary:
The node periodically send getavaaddr and getavaproofs requests to some of its peers. So far only the outbounds were queried, and this diff updates this logic to include the inbounds until the quorum is established. The peers are still selected randomly and upon connection only the outbounds are requested in order to mitigate the possibility for an eclipse attack.

Depends on D11701.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11707