[Cashtab] [Chronik] [Tx Gen] Generate Tx Output
Summary:
As per T2503, this is part of a series of diffs to refactor the transaction generation process within useBCH.

This diff is a standalone utility function which adds tx outputs to the existing transaction builder instance, catering for both one to one and one to many send XEC txs.

The function returns an object containing the transaction builder instance.

Test Plan:
npm test
observe successful execution of generateTxOutput() unit tests

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11719