[Cashtab] [Chronik] [Tx Gen] Sign and build Tx
Summary:
As per T2503, this is part of a series of diffs to refactor the transaction generation process within useBCH.

This diff is a standalone utility function which signs the input UTXOs and converts the transaction builder instance into a raw tx hex ready for broadcast to a node.

This function returns a hex string.

Test Plan:
npm test
observe successful execution of signAndBuildTx() unit tests

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11710