[avalanche] Get rid of the compact proof cleanup log
Summary:
During the compact proof exchange a opy of the radix tree containing the proofs is made, and needs to be cleanup up. This happens when the missing indices are requested, or after some timeout elapsed. In the latter case a message is printed in the log.

The cleanup happens even in the case of an empty tree, or if the tree has been previously cleaned due to a normal proof request. In this case this is a no-op, but the log is still printed which is confusing. D11698 attempted to fix this by avoiding the cleanup and the log if the tree was null, but this is creating a special case just for the sake of fixing the log, as the behavior is otherwise correct.

This diff removes the log which is not very helpful and was introduced to make the associated test more explicit. The test is adapted to use another event indicating the cleanup happened.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11721